% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ca}
\alias{ca}
\title{Information about Counties in California}
\format{The ca data frame contains 55 observations on 4 variables.
\describe{
  \item{City}{City Name}
  \item{Pop10}{Population in Person}
  \item{Pop19}{Population in Person}
  \item{Area(km^2)}{Area in Sqaured Kilometers}
}}
\source{
\url{https://en.wikipedia.org/wiki/List_of_cities_and_towns_in_California}
}
\usage{
ca
}
\description{
This data provides a list of counties in California
with information of name, population in 2010 ans 2019, and land area.
}
\keyword{datasets}
