% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacounty.R
\name{cacounty}
\alias{cacounty}
\title{CACounty}
\usage{
cacounty()
}
\value{
A data frame with all the information about counties in California
}
\description{
This function webscrapes a list of cities and towns in California
with information of name, population in 2010, and land area.
After combining this dataset with another dataset contains
the population for cities and towns in California in 2019,
this function groups information from city level to county level.
}
\examples{
cacounty()
}
