% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ca.R
\name{plot_ca}
\alias{plot_ca}
\title{Plot_CA}
\usage{
plot_ca()
}
\description{
This function plots one scatterplot about the population in 2010
and the area of counties in Californiawith the best fitting line
for the data frame, ca.
}
\examples{
plot_ca()
}
