% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESTinTown.R
\name{RESTinTown}
\alias{RESTinTown}
\title{Resturant_in_Town}
\usage{
RESTinTown(
 
    Yelp_key = "1MHyP9EJgVnXGh7yea0h26Zy2Mf5VQSeyEpl9bvPqgjFmP-au40nu066EI8m4lR4Ky9N7AgTl5xJhtIvTs1ai4txyl4TMlqZp2UjDd-B_-b39DvrjTvhT_t9-a3lXXYx"
)
}
\arguments{
\item{Yelp_key}{User's key for Yelp API}
}
\value{
a list of all information about a county in California
}
\description{
This function requests information about Chinese resturants on Yelp
through Yelp Fusion API by each county in California. In the end,
this function returns a data frame that contains both information
from cacounty and yelp_api, along with new variables of
average number of resturants per person,
average number of resturants per square kilometers.
}
\examples{
RESTinTown("1MHyP9EJgVnXGh7yea0h26Zy2Mf5VQSeyEpl9bvPqgjFmP-au40nu066EI8m4lR4Ky9N7AgTl5xJhtIvTs1ai4txyl4TMlqZp2UjDd-B_-b39DvrjTvhT_t9-a3lXXYx")
}
