% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yelp_api.R
\name{yelp_api}
\alias{yelp_api}
\title{Yelp Fusion API}
\usage{
yelp_api(
  county = "San Deigo",
 
    Yelp_key = "1MHyP9EJgVnXGh7yea0h26Zy2Mf5VQSeyEpl9bvPqgjFmP-au40nu066EI8m4lR4Ky9N7AgTl5xJhtIvTs1ai4txyl4TMlqZp2UjDd-B_-b39DvrjTvhT_t9-a3lXXYx"
)
}
\arguments{
\item{county}{the county name in California}

\item{Yelp_key}{User's key for Yelp API}
}
\value{
information from Yelp about the Chinese resturants in the County
}
\description{
This function requests information about Chinese resturants
in one California county on Yelp through Yelp Fusion API.
After some analysis, this function returns a data frame with
information about the average rating and average review counts
for Chinese resturant in the county, along with the name,
rating, and review counts of the best-rated Chinese resturant.
}
\examples{
yelp_api("San Diego", "1MHyP9EJgVnXGh7yea0h26Zy2Mf5VQSeyEpl9bvPqgjFmP-au40nu066EI8m4lR4Ky9N7AgTl5xJhtIvTs1ai4txyl4TMlqZp2UjDd-B_-b39DvrjTvhT_t9-a3lXXYx")
}
